mail2rag/requirements.txt
certifi
imapclient
jinja2
pdf2image
pdfplumber
Pillow
pytesseract
requests
urllib3
